refactor: move hardware logic into backends
This is pretty awful, but it's the only way to make this work with
simulation:
- we want the ot3api to be the same between simulating and not
simulating
- we want the backend to behave reasonably between simulating and not
simulating
- we want to keep some nice sanity checks

Even if we move the sanity checks down into the backend, we'd need to
tell them what's sane - the backends don't typically look at the
configuration for instruments.

This setup is a little ugly but it's really fundamentally about moving a
pure function around, and injecting some context knowledge into a
simulator as one occasionally has to do.
