+++ api/src/opentrons/hardware_control/backends/flex_protocol.py
@@ -188,6 +188,7 @@ class FlexBackend(Protocol):
     async def gripper_grip_jaw(
         self,
         duty_cycle: float,
+        expected_displacement: float,
         stop_condition: HWStopCondition = HWStopCondition.none,
         stay_engaged: bool = True,
     ) -> None:
@@ -395,4 +396,16 @@ class FlexBackend(Protocol):
         ...
 
     def add_estop_callback(self, cb: HardwareEventHandler) -> HardwareEventUnsubscriber:
+        ...
+
+    def check_gripper_position_within_bounds(
+        self,
+        expected_grip_width: float,
+        grip_width_uncertainty_wider: float,
+        grip_width_uncertainty_narrower: float,
+        jaw_width: float,
+        max_allowed_grip_error: float,
+        hard_limit_lower: float,
+        hard_limit_upper: float,
+    ) -> None:
         ...
