+++ api/src/opentrons/hardware_control/backends/ot3simulator.py
@@ -400,16 +400,19 @@ class OT3Simulator(FlexBackend):
     async def gripper_grip_jaw(
         self,
         duty_cycle: float,
+        expected_displacement: float,
         stop_condition: HWStopCondition = HWStopCondition.none,
         stay_engaged: bool = True,
     ) -> None:
         """Move gripper inward."""
         self._sim_jaw_state = GripperJawState.GRIPPING
+        self._encoder_position[Axis.G] = expected_displacement
 
     @ensure_yield
     async def gripper_home_jaw(self, duty_cycle: float) -> None:
         """Move gripper outward."""
         self._motor_status[Axis.G] = MotorStatus(True, True)
+        self._encoder_position[Axis.G] = self._get_home_position()[Axis.G]
         self._sim_jaw_state = GripperJawState.HOMED_READY
 
     @ensure_yield
@@ -778,3 +781,17 @@ class OT3Simulator(FlexBackend):
 
     def add_estop_callback(self, cb: HardwareEventHandler) -> HardwareEventUnsubscriber:
         return lambda: None
+
+    def check_gripper_position_within_bounds(
+        self,
+        expected_grip_width: float,
+        grip_width_uncertainty_wider: float,
+        grip_width_uncertainty_narrower: float,
+        jaw_width: float,
+        max_allowed_grip_error: float,
+        hard_limit_lower: float,
+        hard_limit_upper: float,
+    ) -> None:
+        # This is a (pretty bad) simulation of the gripper actually gripping something,
+        # but it should work.
+        self._encoder_position[Axis.G] = (hard_limit_upper - jaw_width) / 2
