+++ shared-data/pipette/definitions/2/liquid/single_channel/p50/default/3_5.json
@@ -20,50 +20,50 @@
       "aspirate": {
         "default": {
           "1": [
-            [0.462, 0.5646, 0.0415],
-            [0.648, 0.3716, 0.1307],
-            [1.032, 0.2742, 0.1938],
-            [1.37, 0.1499, 0.3221],
-            [2.014, 0.1044, 0.3845],
-            [2.772, 0.0432, 0.5076],
-            [3.05, -0.0809, 0.8517],
-            [3.4, 0.0256, 0.5268],
-            [3.962, 0.0612, 0.4057],
-            [4.438, 0.0572, 0.4217],
-            [5.164, 0.018, 0.5955],
-            [5.966, 0.0095, 0.6393],
-            [7.38, 0.0075, 0.6514],
-            [9.128, 0.0049, 0.6705],
-            [10.16, 0.0033, 0.6854],
-            [13.812, 0.0024, 0.6948],
-            [27.204, 0.0008, 0.7165],
-            [50.614, 0.0002, 0.7328],
-            [53.046, -0.0005, 0.7676]
+            [0.3100,0.5910,0.0197],
+            [0.3900,0.2586,0.1227],
+            [0.8600,0.3697,0.0794],
+            [1.2900,0.2310,0.1987],
+            [1.9300,0.1144,0.3491],
+            [2.7000,0.0536,0.4664],
+            [2.9500,-0.1041,0.8923],
+            [3.2800,0.0216,0.5214],
+            [3.7600,0.0480,0.4349],
+            [4.3800,0.0830,0.3032],
+            [5.0800,0.0153,0.5996],
+            [5.9000,0.0136,0.6083],
+            [7.2900,0.0070,0.6474],
+            [9.0400,0.0059,0.6551],
+            [10.0800,0.0045,0.6682],
+            [13.7400,0.0029,0.6842],
+            [27.1500,0.0010,0.7104],
+            [50.4800,0.0002,0.7319],
+            [52.8900,-0.0006,0.7703]
           ]
         }
       },
       "dispense": {
         "default": {
           "1": [
-            [0.462, 0.5646, 0.0415],
-            [0.648, 0.3716, 0.1307],
-            [1.032, 0.2742, 0.1938],
-            [1.37, 0.1499, 0.3221],
-            [2.014, 0.1044, 0.3845],
-            [2.772, 0.0432, 0.5076],
-            [3.05, -0.0809, 0.8517],
-            [3.4, 0.0256, 0.5268],
-            [3.962, 0.0612, 0.4057],
-            [4.438, 0.0572, 0.4217],
-            [5.164, 0.018, 0.5955],
-            [5.966, 0.0095, 0.6393],
-            [7.38, 0.0075, 0.6514],
-            [9.128, 0.0049, 0.6705],
-            [10.16, 0.0033, 0.6854],
-            [13.812, 0.0024, 0.6948],
-            [27.204, 0.0008, 0.7165],
-            [50.614, 0.0002, 0.7328],
-            [53.046, -0.0005, 0.7676]
+            [0.3100,0.5910,0.0197],
+            [0.3900,0.2586,0.1227],
+            [0.8600,0.3697,0.0794],
+            [1.2900,0.2310,0.1987],
+            [1.9300,0.1144,0.3491],
+            [2.7000,0.0536,0.4664],
+            [2.9500,-0.1041,0.8923],
+            [3.2800,0.0216,0.5214],
+            [3.7600,0.0480,0.4349],
+            [4.3800,0.0830,0.3032],
+            [5.0800,0.0153,0.5996],
+            [5.9000,0.0136,0.6083],
+            [7.2900,0.0070,0.6474],
+            [9.0400,0.0059,0.6551],
+            [10.0800,0.0045,0.6682],
+            [13.7400,0.0029,0.6842],
+            [27.1500,0.0010,0.7104],
+            [50.4800,0.0002,0.7319],
+            [52.8900,-0.0006,0.7703]
           ]
         }
       },
