+++ shared-data/pipette/definitions/2/liquid/single_channel/p50/lowVolumeDefault/3_5.json
@@ -20,46 +20,48 @@
       "aspirate": {
         "default": {
           "1": [
-            [0.11, 0.207815, 0.040201],
-            [0.65, 0.43933, 0.014735],
-            [1.04, 0.256666, 0.133466],
-            [1.67, 0.147126, 0.247388],
-            [2.45, 0.078774, 0.361536],
-            [2.89, 0.042387, 0.450684],
-            [3.2, 0.014781, 0.530464],
-            [3.79, 0.071819, 0.347944],
-            [4.22, 0.051592, 0.424605],
-            [4.93, 0.021219, 0.552775],
-            [5.81, 0.023461, 0.541725],
-            [7.21, 0.008959, 0.625982],
-            [8.93, 0.005456, 0.651235],
-            [10.0, 0.007108, 0.636489],
-            [13.61, 0.002591, 0.681656],
-            [26.99, 0.001163, 0.701094],
-            [45.25, 0.000207, 0.726887]
+            [0.3000,0.4590,0.0586],
+            [0.4700,0.4300,0.0674],
+            [0.9000,0.3404,0.1095],
+            [1.2600,0.1925,0.2425],
+            [1.9500,0.1314,0.3195],
+            [2.7600,0.0604,0.4580],
+            [2.9500,-0.2085,1.2002],
+            [3.3300,0.0425,0.4597],
+            [3.8700,0.0592,0.4040],
+            [4.3100,0.0518,0.4327],
+            [5.0700,0.0264,0.5424],
+            [5.9300,0.0186,0.5818],
+            [7.3400,0.0078,0.6458],
+            [9.0800,0.0050,0.6664],
+            [10.0900,0.0022,0.6918],
+            [13.7400,0.0027,0.6868],
+            [27.1300,0.0009,0.7109],
+            [45.4300,-0.0038,0.8391]
           ]
         }
       },
       "dispense": {
         "default": {
           "1": [
-            [0.11, 0.207815, 0.040201],
-            [0.65, 0.43933, 0.014735],
-            [1.04, 0.256666, 0.133466],
-            [1.67, 0.147126, 0.247388],
-            [2.45, 0.078774, 0.361536],
-            [2.89, 0.042387, 0.450684],
-            [3.2, 0.014781, 0.530464],
-            [3.79, 0.071819, 0.347944],
-            [4.22, 0.051592, 0.424605],
-            [4.93, 0.021219, 0.552775],
-            [5.81, 0.023461, 0.541725],
-            [7.21, 0.008959, 0.625982],
-            [8.93, 0.005456, 0.651235],
-            [10.0, 0.007108, 0.636489],
-            [13.61, 0.002591, 0.681656],
-            [26.99, 0.001163, 0.701094],
-            [45.25, 0.000207, 0.726887]
+            [0.3000,0.4590,0.0586],
+            [0.4700,0.4300,0.0674],
+            [0.9000,0.3404,0.1095],
+            [1.2600,0.1925,0.2425],
+            [1.9500,0.1314,0.3195],
+            [2.7600,0.0604,0.4580],
+            [2.9500,-0.2085,1.2002],
+            [3.3300,0.0425,0.4597],
+            [3.8700,0.0592,0.4040],
+            [4.3100,0.0518,0.4327],
+            [5.0700,0.0264,0.5424],
+            [5.9300,0.0186,0.5818],
+            [7.3400,0.0078,0.6458],
+            [9.0800,0.0050,0.6664],
+            [10.0900,0.0022,0.6918],
+            [13.7400,0.0027,0.6868],
+            [27.1300,0.0009,0.7109],
+            [45.4300,-0.0038,0.8391]
           ]
         }
       },
